+++ cookbooks/video/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file is used to list changes made in each version of video.
 
+## 0.3.0:
+
+* recipes/default - now includes miro and mmc install management
+* recipes/miro    - installs Miro media player
+* recipes/mmc     - installs Mobile Media Converter
+
 ## 0.2.0:
 
 * recipe/default - now checks for section presence in box profile's apps list before proceed